feat: amends spacing between sections and renames section style modules
