+++ styles/BioSection.module.sass
@@ -9,7 +9,7 @@
     padding-top: 2rem
     
 .bioImage
-    height: 450px
+    height: 500px
     border-radius: 1rem
     background-color: #2D2F3A
