+++ styles/CodeSection.module.sass
@@ -1,6 +1,3 @@
-.codeSection
-    margin: 4rem 0
-    
 .codeRepoGrid
     display: grid
     grid-template-columns: 1fr 1fr 1fr 1fr
