+++ styles/Footer.module.sass
@@ -1,6 +1,6 @@
 .footer
     padding: 1rem
-    padding: 0 0 1rem
+    padding: 0 0 3rem
 
     hr
         margin-bottom: 1rem
