+++ styles/ProjectsSection.module.sass
@@ -2,7 +2,8 @@
     background-color: #2D2F3A
     border-radius: 1rem
     padding: 2rem
-    margin: 1rem -1rem
+    margin-right: -1rem
+    margin-left: -1rem
 
 .projectGrid
     display: grid
