+++ styles/globals.sass
@@ -14,7 +14,7 @@ main
     margin: 0 auto
 
 section
-    margin-bottom: 4rem
+    margin: 3rem 0 4rem
 
 h1
     font-size: 2.4rem
